crypto_viz_backend/requirements.txt
# Django et extensions
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Documentation API (Swagger/OpenAPI)
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1

# Base de données
psycopg2-binary==2.9.11

# Kafka
kafka-python==2.0.2

# Django Channels (optionnel pour WebSocket)
channels==4.0.0
channels-redis==4.1.0

# Cache Redis
django-redis==5.4.0
redis==5.0.1
hiredis==2.3.2

# Utilitaires
python-dotenv==1.0.0

# Monitoring et logs
gunicorn==21.2.0
prometheus-client==0.19.0
django-prometheus==2.3.1

# Static files serving
whitenoise==6.11.0
